requirements.txt
streamlit
pandas
numpy
langchain
langchain-groq
langgraph
plotly
scipy
python-dotenv
pydantic
